g_io_channel_read_chars takes a gsize.
2003-08-31  Manish Singh  <yosh@gimp.org>

        * plug-ins/helpbrowser/dialog.c (io_handler): g_io_channel_read_chars
        takes a gsize.

+++ plug-ins/helpbrowser/dialog.c
@@ -571,7 +571,7 @@ io_handler (GIOChannel   *io,
 {
   HtmlStream *stream;
   gchar       buffer[8192];
-  guint       bytes;
+  gsize       bytes;
 
   stream = (HtmlStream *) data;
